refactor(lv7): 조회 관련 메서드 재활용을 위한 Finder 분리
 -공용 default 메서드 삭제
