+++ src/main/java/io/github/seonrizee/scheduler/repository/CommentRepository.java
@@ -6,5 +6,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
+
     List<Comment> findAllBySchedule(Schedule schedule);
 }
