+++ src/main/java/io/github/seonrizee/scheduler/service/CommentService.java
@@ -3,16 +3,11 @@ package io.github.seonrizee.scheduler.service;
 import io.github.seonrizee.scheduler.dto.request.CommentCreateRequest;
 import io.github.seonrizee.scheduler.dto.request.CommentUpdateRequest;
 import io.github.seonrizee.scheduler.dto.response.CommentDetailResponse;
-import io.github.seonrizee.scheduler.dto.response.CommentListResponse;
 import io.github.seonrizee.scheduler.entity.User;
 
 public interface CommentService {
 
     CommentDetailResponse createComment(Long scheduleId, CommentCreateRequest requestDto, User user);
-
-    CommentDetailResponse getComment(Long commentId);
-
-    CommentListResponse getCommentsWithSchedule(Long scheduleId);
 
     CommentDetailResponse updateComment(Long commentId, CommentUpdateRequest requestDto, User user);
