+++ src/main/java/io/github/seonrizee/scheduler/service/ScheduleService.java
@@ -3,17 +3,12 @@ package io.github.seonrizee.scheduler.service;
 import io.github.seonrizee.scheduler.dto.request.ScheduleCreateRequest;
 import io.github.seonrizee.scheduler.dto.request.ScheduleUpdateRequest;
 import io.github.seonrizee.scheduler.dto.response.ScheduleDetailResponse;
-import io.github.seonrizee.scheduler.dto.response.ScheduleListResponse;
 import io.github.seonrizee.scheduler.entity.User;
 
 public interface ScheduleService {
 
     ScheduleDetailResponse createSchedule(ScheduleCreateRequest requestDto, User user);
-
-    ScheduleDetailResponse findScheduleById(Long scheduleId);
-
-    ScheduleListResponse getSchedules();
-
+    
     ScheduleDetailResponse updateSchedule(Long scheduleId, ScheduleUpdateRequest requestDto, User user);
 
     void deleteSchedule(Long scheduleId, User user);
